requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
dataclasses==0.8
docker==2.6.1
docker-compose==1.18.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.1.0
idna==3.4
importlib-metadata==4.8.3
joblib==1.1.1
jsonschema==2.5.1
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.19.5
oauthlib==3.2.2
opt-einsum==3.3.0
paho-mqtt==1.6.1
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
PySocks==1.6.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2023.3
PyYAML==3.13
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==3.1.0
tk==0.1.0
typing-extensions==3.7.4.3
urllib3==1.26.16
websocket-client==0.47.0
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
